+++ frontend/docs/pages/home/streaming.mdx
@@ -23,7 +23,7 @@ You can stream data out of a task run by using the `put_stream` (or equivalent) 
 </Tabs.Tab>
 <Tabs.Tab title="Go">
 
-<Snippet src={snips.go.streaming.worker} block="streaming" />
+<Snippet src={snips.go.streaming.shared.task} block="streaming" />
 
 </Tabs.Tab>
 
@@ -47,7 +47,7 @@ You can easily consume stream events by using the stream method on the workflow 
 </Tabs.Tab>
 <Tabs.Tab title="Go">
 
-<Snippet src={snips.go.streaming.async_stream} block="consume" />
+<Snippet src={snips.go.streaming.consumer.main} block="consume" />
 
 </Tabs.Tab>
 
@@ -93,7 +93,7 @@ For example, with NextJS backend-as-frontend, you'd do the following:
 
 For example, with Go's built-in HTTP server, you'd do the following:
 
-<Snippet src={snips.go.streaming.http_proxy} block="http_proxy" />
+<Snippet src={snips.go.streaming.server.main} block="server" />
 
 </Tabs.Tab>
